Refactor paths

+++ Stratum CNC/projects/ProjectsStore.swift
@@ -21,18 +21,10 @@ final class ProjectsStore: ObservableObject {
     @Published var activeProject: Project?
     @Published var activeProjectModel: ProjectModel?
 
-    let paths: ProjectPaths
-
     private let encoder: JSONEncoder
     private let decoder: JSONDecoder
 
     init() {
-        do {
-            self.paths = try ProjectPaths()
-        } catch {
-            fatalError("Unable to initialize project storage: \(error)")
-        }
-
         self.encoder = JSONEncoder()
         self.decoder = JSONDecoder()
 
@@ -46,13 +38,13 @@ final class ProjectsStore: ObservableObject {
     // MARK: - Loading
 
     func loadProjects() {
-        guard FileManager.default.fileExists(atPath: paths.projectsIndexFile.path) else {
+        guard FileManager.default.fileExists(atPath: ProjectPaths.projectsIndexFile.path) else {
             projects = []
             return
         }
 
         do {
-            let data = try Data(contentsOf: paths.projectsIndexFile)
+            let data = try Data(contentsOf: ProjectPaths.projectsIndexFile)
             projects = try decoder.decode([Project].self, from: data)
 
             projects.sort {
@@ -65,12 +57,18 @@ final class ProjectsStore: ObservableObject {
     }
 
     func open(_ project: Project) {
+        guard let paths = try? ProjectPaths(project: project) else {
+            return
+        }
         activeProject = project
         activeProjectModel = ProjectModel(project: project, paths: paths)
     }
 
     func openZombieProject() {
         let project = Project(name: "Untitled")
+        guard let paths = try? ProjectPaths(project: project) else {
+            return
+        }
         activeProject = project
         activeProjectModel = ProjectModel(project: project, paths: paths)
     }
@@ -84,52 +82,56 @@ final class ProjectsStore: ObservableObject {
     // MARK: - Creation
 
     @discardableResult
-    func createProject(name: String) throws -> (Project, ProjectData) {
-        // Forget the previous project
-        activeProject = nil
-        activeProjectModel = nil
-
-        let project = Project(name: name)
-        // Create project data
-        let stock = StockMaterial(name: "Aluminum", material: .aluminum, geometry: .rectangular(width: 150, height: 25, depth: 5))
-        let projectData = ProjectData(stock: stock, isStockVisible: true, assets: nil)
+    func createProject(name: String) throws -> Project {
 
         // Check if project already exist
         guard projects.first(where: { $0.name.lowercased() == name.lowercased() }) == nil else {
             throw ProjectError.projectAlreadyExists
         }
 
-        // Create supporting files
-        try FileManager.default.createDirectory(at: paths.projectDirectory(for: project), withIntermediateDirectories: true)
-        try FileManager.default.createDirectory(at: paths.assetsDirectory(for: project), withIntermediateDirectories: true)
+        let project = Project(name: name)
 
+        guard let paths = try? ProjectPaths(project: project) else {
+            throw ProjectError.projectNotFound
+        }
+
+        let lastUsedGeometry = StockGeometry.rectangular(width: 150, height: 25, depth: 5)
+        let lastStockVisibility = true
+        let lastStockUsed = StockMaterial(name: "Aluminum", material: .aluminum, geometry: lastUsedGeometry)
+        let projectData = ProjectData(stock: lastStockUsed, isStockVisible: lastStockVisibility, assets: nil)
+
+        // Create supporting directories and files
+        try FileManager.default.createDirectory(at: paths.projectDirectory, withIntermediateDirectories: true)
+        try FileManager.default.createDirectory(at: paths.assetsDirectory, withIntermediateDirectories: true)
         try saveProjectMetadata(projectData, in: project)
 
         projects.insert(project, at: 0)
 
         try saveIndex()
 
-        return (project, projectData)
+        return project
     }
 
     // MARK: - Updating
 
-//    func update(_ project: Project) throws {
-//        var updatedProject = project
-//        updatedProject.modifiedAt = .now
-//
-//        if let index = projects.firstIndex(where: { $0.id == project.id }) {
-//            projects[index] = updatedProject
-//        }
-//
-//        try saveProjectMetadata(updatedProject)
-//        try saveIndex()
-//    }
+    func update(_ project: Project) throws {
+        var updatedProject = project
+        updatedProject.modifiedAt = .now
+
+        if let index = projects.firstIndex(where: { $0.id == project.id }) {
+            projects[index] = updatedProject
+        }
+
+        try saveIndex()
+    }
 
     // MARK: - Delete
 
     func delete(_ project: Project) throws {
-        let directory = paths.projectDirectory(for: project)
+        guard let paths = try? ProjectPaths(project: project) else {
+            return
+        }
+        let directory = paths.projectDirectory
 
         if FileManager.default.fileExists(atPath: directory.path) {
             try FileManager.default.removeItem(at: directory)
@@ -145,18 +147,24 @@ final class ProjectsStore: ObservableObject {
     // MARK: - Preview
 
     func previewURL(for project: Project) -> URL {
-        paths.preview(for: project)
+        guard let paths = try? ProjectPaths(project: project) else {
+            return URL(fileURLWithPath: "")
+        }
+        return paths.preview
     }
 
     // MARK: - Persistence
 
     private func saveProjectMetadata(_ projectData: ProjectData, in project: Project) throws {
+        guard let paths = try? ProjectPaths(project: project) else {
+            return
+        }
         let data = try encoder.encode(projectData)
-        try data.write(to: paths.projectMetadata(for: project), options: .atomic)
+        try data.write(to: paths.projectMetadata, options: .atomic)
     }
 
     private func saveIndex() throws {
         let data = try encoder.encode(projects)
-        try data.write(to: paths.projectsIndexFile, options: .atomic)
+        try data.write(to: ProjectPaths.projectsIndexFile, options: .atomic)
     }
 }
